update: updating structure folder for routing
