+++ app/login.tsx
@@ -1,0 +1,7 @@
+import LoginScreen from "./screens/login";
+
+const Login = () => {
+  return <LoginScreen />;
+};
+
+export default Login
