+++ app/register.tsx
@@ -1,0 +1,7 @@
+import RegisterScreen from "./screens/register";
+
+const Register = () => {
+  return <RegisterScreen />;
+};
+
+export default Register;
