+++ app/screens/register.tsx
@@ -1,7 +1,6 @@
-// app/register/index.tsx
 import { View, Text, StyleSheet } from 'react-native';
 
-export default function Register() {
+const RegisterScreen = () => {
   return (
     <View style={styles.container}>
       <Text style={styles.text}>Ini Halaman Register</Text>
@@ -13,3 +12,5 @@ const styles = StyleSheet.create({
   container: { flex: 1, justifyContent: 'center', alignItems: 'center' },
   text: { fontSize: 20 },
 });
+
+export default RegisterScreen
